+++ backend/canisters/user/impl/src/lifecycle/init.rs
@@ -26,7 +26,7 @@ fn init(args: Args) {
 
     init_state(env, data, args.wasm_version);
 
-    openchat_bot::send_welcome_messages();
+    openchat_bot::send_welcome_message();
 
     info!(version = %args.wasm_version, "Initialization complete");
 }