+++ backend/canisters/user_index/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Share Diamond membership payment with referrer ([#3452](https://github.com/open-ic/open-chat/pull/3452))
+- Send OpenChatBot welcome messages from the UserIndex ([#3478](https://github.com/open-ic/open-chat/pull/3478))
 
 ### Removed